train/pyproject.toml
[project]
name = "train"
requires-python = ">=3.10, <3.12"
dependencies = [
  "tensorflow (==2.15.0)",
  "matplotlib (>=3.10.0,<4.0.0)",
  "tensorflowjs (>=4.22.0,<5.0.0)",
  "tensorflow-decision-forests (==1.8.1)",
  "pillow (>=11.1.0,<12.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
